requirements.txt
datetime
FAISS-cpu
langchain
langchain-community
langchain-core
langchain-ollama
langsmith
numpy
simsimd
openpyxl
pandas
pypdf
PyPDF2
python-docx
requests
torch
pytesseract
pdf2image
pdfplumber
python-decouple